fix: restructure project folders and improve code readability
