+++ src/app/dashboard/main/page.tsx
@@ -1,4 +1,4 @@
-import { SimpleWidget } from "@/components";
+import { WidgetsGrid } from "@/components";
 
 
 export const metadata = {
@@ -11,9 +11,7 @@ export default function MainPage() {
     <div className="text-black p-2">
       <h1 className="mt-2 text-3xl">Dashboard</h1>
       <span className="text-xl">Informacion general</span>
-      <div className="flex flex-wrap p-2 items-center justify-center">
-          <SimpleWidget />
-      </div>
+      <WidgetsGrid />
     </div>
   );
 }
