+++ src/components/index.ts
@@ -1,4 +1,5 @@
 
-export { Sidebar } from './Sidebar';
-export { SidebarMenuItem } from './SidebarMenuItem';
-export { SimpleWidget } from './SimpleWidget';
+export { Sidebar } from './sidebar/Sidebar';
+export { SidebarMenuItem } from './sidebar/SidebarMenuItem';
+export { SimpleWidget } from './dashboard/SimpleWidget';
+export { WidgetsGrid } from './dashboard/WidgetsGrid';
